Add functions that help debugging lightmaps with UVs
The new function generateCombinedLightmap is used to create a lightmap
that contains the lighting informations of the meshes given to the
function.
The resulting lightmap can then be drawn with graphics debugger to the
screen.

The function debugUVs generate a transparent canvas of the given UVs.

Drawing this canvas at the same position the lightmaps is draw with the
debugger make it easy to see if UVs are correct and math the lightmap
infos.
It also help to see if multiple meshes have their lighting informations
at the same placing causing graphics atifacts.