Fix rotated lights not generating light.
Some lights rotated on the y axis didn't generates any light.

As the visibility for this light was empty their was a bug in the way we
gather the depth from the scene.

I fixed this by changing the cube projection matrices to legacy
projection matrices, and gathering the depth by just dividing it by w.

Then some depth acne appeared, caused by the streched sides maps.
Now the sides maps are rendered fully.

This was tested on multiple scenes, some that was already correct and
some bugged. The correct ones remain correct and clean, while the bugged
one looks good now.